Implemented end game. Needs bug fix/balance. Made terrain generator, doesn't destroy anything, can be used with loop to get terrain from file

+++ scripts/move_collision.gml
@@ -23,8 +23,17 @@ if (object_index == oPlayer)
             //var cell_next = oGame.map_update[argument0, argument1];
             if (cell_next.object_index != oSuperwall)
             {
-                oGame.map[argument0, argument1] = instance_create(argument0, argument1, cell_next.dead);
-                oGame.map_update[argument0, argument1] = oGame.map[argument0, argument1]; // TODO: this code makes the game crash when you kill yourself for some reason?
+                
+                if (instance_number(oTree) == 1)
+                {
+                    oGame.map[argument0, argument1] = instance_create(argument0, argument1, oExit);
+                    oGame.map_update[argument0, argument1] = oGame.map[argument0, argument1];
+                }
+                else
+                {
+                    oGame.map[argument0, argument1] = instance_create(argument0, argument1, cell_next.dead);
+                    oGame.map_update[argument0, argument1] = oGame.map[argument0, argument1]; // TODO: this code makes the game crash when you kill yourself for some reason?
+                }
                 if (!audio_is_playing(cell_next.dead_sound))
                 {
                     audio_play_sound(cell_next.dead_sound, 1, false);
